fix(sockets): widen UmaskGuard mask to u32 for Linux Mode::from
rustix's Mode only implements From<u32>; Linux mode_t is u32 while
the previous u16 parameter compiled on Darwin but broke the ubuntu
build (E0277). The whole UmaskGuard block is already gated to
non-macOS Unix, so u32 matches the platform mode_t with no semantic
change.

+++ crates/cairn-core/src/sockets.rs
@@ -212,7 +212,7 @@ struct UmaskGuard {
 
 #[cfg(all(unix, not(target_os = "macos")))]
 impl UmaskGuard {
-    fn set(mask: u16) -> Self {
+    fn set(mask: u32) -> Self {
         let lock = UMASK_LOCK
             .lock()
             .unwrap_or_else(|poisoned| poisoned.into_inner());
